+++ internal/sync/upsert.go
@@ -622,7 +622,7 @@ func upsertIxLans(ctx context.Context, tx *ent.Tx, items []peeringdb.IxLan) ([]i
 				SetDot1qSupport(il.Dot1QSupport).
 				SetNillableRsAsn(il.RSASN).
 				SetNillableArpSponge(il.ARPSponge).
-				SetIxfIxpMemberListURL(il.IXFIXPMemberListURL). // auth-gated field
+				SetNillableIxfIxpMemberListURL(il.IXFIXPMemberListURL). // auth-gated field
 				SetIxfIxpMemberListURLVisible(il.IXFIXPMemberListURLVisible).
 				SetIxfIxpImportEnabled(il.IXFIXPImportEnabled).
 				SetCreated(il.Created).